pyproject.toml
[tool.poetry]
name = "django-meilisearch"
version = "0.1.0"
description = ""
authors = ["Lucas Rocha <lucash.rocha@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4"
meilisearch = "^0.31.4"
pydantic = "^2.8.2"
djantic2 = "^1.0.4"
alive-progress = "^3.1.5"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-django = "^4.8.0"
pytest-cov = "^5.0.0"
black = "^24.4.2"
taskipy = "^1.12.2"
pylint = "^3.1.0"
mypy = "^1.11.2"
django-stubs = {extras = ["compatible-mypy"], version = "^5.0.4"}
pylint-django = "^2.5.5"

[tool.taskipy.tasks]
serve = "python src/manage.py runserver"
format = "black src"
lint = "pylint src"
check = "mypy src"
test = "pytest -v"
cov = "coverage run -m pytest && coverage report"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "mysite.settings"
python_files = ["test_*", "*_test", "tests/*"]
pythonpath = ["src"]

[tool.mypy]
ignore_missing_imports = true

[tool.pylint]
load-plugins = ["pylint_django"]
django-settings-module = "mysite.settings"
ignore = ["migrations"]

[tool.poetry.scripts]
check = "scripts.check:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

